Added crypto domain
This domain appears in an email spam campaign. The email has gibberish subject and title, and seems to propagate when victim opens the attached `xhtml` file. This file has JS code set on execute on page open and redirects user to `btcminings.site`. The actual url is encoded as base64.