Remove AWS kinesis from ads list
AWS kinesis is a data stream processing platform in the amazon cloud, it
is not an ad server.

Furthermore, it is a regional service and the ads.txt list only included
one region (us-east-1), so it wasn't effectively blocking this service
anyways. In my home, video cameras were trying and failing to reach
kinesis.us-east-1, but then retrying kinesis.us-west-2 and succeeding to
use kinesis.

see https://aws.amazon.com/kinesis/